move facility charts to facility dashboard
