+++ src/app/app.module.ts
@@ -50,9 +50,9 @@ import { ImportPredictorsComponent } from './utility/predictor-data/import-predi
 import { EnergyUnitDropdownComponent } from './utility/meter-grouping/energy-unit-dropdown/energy-unit-dropdown.component';
 import { SettingsLabelPipe } from './shared/helper-pipes/settings-label.pipe';
 import { VisualizationComponent } from './utility/visualization/visualization.component';
-import { FacilityBarChartComponent } from './utility/visualization/facility-bar-chart/facility-bar-chart.component';
-import { FacilityHeatMapComponent } from './utility/visualization/facility-heat-map/facility-heat-map.component';
-import { FacilityStackedAreaChartComponent } from './utility/visualization/facility-stacked-area-chart/facility-stacked-area-chart.component';
+import { FacilityBarChartComponent } from './dashboard/facility-overview/facility-bar-chart/facility-bar-chart.component';
+import { FacilityHeatMapComponent } from './dashboard/facility-overview/facility-heat-map/facility-heat-map.component';
+import { FacilityStackedAreaChartComponent } from './dashboard/facility-overview/facility-stacked-area-chart/facility-stacked-area-chart.component';
 import { FacilityOverviewComponent } from './dashboard/facility-overview/facility-overview.component';
 import { AccountOverviewComponent } from './dashboard/account-overview/account-overview.component';
 import { FacilitiesTableComponent } from './dashboard/account-overview/facilities-table/facilities-table.component';
