+++ src/app/dashboard/facility-overview/facility-bar-chart/facility-bar-chart.component.spec.ts
