+++ src/app/dashboard/facility-overview/facility-bar-chart/facility-bar-chart.component.ts
@@ -2,50 +2,77 @@
 import { Component, ElementRef, OnInit, ViewChild } from '@angular/core';
 import { PlotlyService } from 'angular-plotly.js';
 import { Subscription } from 'rxjs';
+import { FacilitydbService } from 'src/app/indexedDB/facility-db.service';
 import { UtilityMeterdbService } from 'src/app/indexedDB/utilityMeter-db.service';
 import { UtilityMeterDatadbService } from 'src/app/indexedDB/utilityMeterData-db.service';
-import { IdbUtilityMeter } from 'src/app/models/idb';
-import { CalanderizationService } from '../../calanderization/calanderization.service';
-import { VisualizationService } from '../visualization.service';
+import { IdbFacility, IdbUtilityMeter, IdbUtilityMeterData } from 'src/app/models/idb';
+import { VisualizationService } from '../../../utility/visualization/visualization.service';
 
 @Component({
-  selector: 'app-facility-stacked-area-chart',
-  templateUrl: './facility-stacked-area-chart.component.html',
-  styleUrls: ['./facility-stacked-area-chart.component.css']
+  selector: 'app-facility-bar-chart',
+  templateUrl: './facility-bar-chart.component.html',
+  styleUrls: ['./facility-bar-chart.component.css']
 })
-export class FacilityStackedAreaChartComponent implements OnInit {
+export class FacilityBarChartComponent implements OnInit {
+  /*Grouped bar chart
+    Group by utility
+    Bars are energy usage and cost by time
+    Month or year (user selection)
+  */
 
-  @ViewChild('stackedAreaChart', { static: false }) stackedAreaChart: ElementRef;
-  facilityMetersSub: Subscription;
-  facilityMeters: Array<IdbUtilityMeter>;
+
+  @ViewChild('utilityBarChart', { static: false }) utilityBarChart: ElementRef;
+
+
+
   electricityData: Array<{ time: string, energyUse: number, energyCost: number }>;
   naturalGasData: Array<{ time: string, energyUse: number, energyCost: number }>;
   otherFuelsData: Array<{ time: string, energyUse: number, energyCost: number }>;
   waterData: Array<{ time: string, energyUse: number, energyCost: number }>;
   wasteWaterData: Array<{ time: string, energyUse: number, energyCost: number }>;
   otherUtilityData: Array<{ time: string, energyUse: number, energyCost: number }>;
+  facilityMeters: Array<IdbUtilityMeter>;
   sumByMonth: boolean = false;
   removeIncompleteYears: boolean = true;
+
+  selectedFacility: IdbFacility;
+  selectedFacilitySub: Subscription;
+  accountMeterDataSub: Subscription;
+  accountMeters: Array<IdbUtilityMeter>;
+  accountMetersSub: Subscription;
   constructor(private plotlyService: PlotlyService, private utilityMeterDbService: UtilityMeterdbService,
-    private utilityMeterDataDbService: UtilityMeterDatadbService, private vizualizationService: VisualizationService) { }
+    private utilityMeterDataDbService: UtilityMeterDatadbService, private vizualizationService: VisualizationService,
+    private facilityDbService: FacilitydbService) { }
 
   ngOnInit(): void {
-    this.facilityMetersSub = this.utilityMeterDataDbService.facilityMeterData.subscribe(() => {
-      let facilityMeters: Array<IdbUtilityMeter> = this.utilityMeterDbService.facilityMeters.getValue();
-      this.facilityMeters = JSON.parse(JSON.stringify(facilityMeters))
+    this.accountMetersSub = this.utilityMeterDbService.accountMeters.subscribe(accountMeters => {
+      this.accountMeters = accountMeters;
       this.setUtilityData();
+    });
+    this.selectedFacilitySub = this.utilityMeterDbService.facilityMeters.subscribe(facilityMeters => {
+      this.facilityMeters = JSON.parse(JSON.stringify(facilityMeters));
+      this.setUtilityData();
+    });
+
+    this.accountMeterDataSub = this.utilityMeterDataDbService.accountMeterData.subscribe(accountMeterData => {
+      if (accountMeterData && accountMeterData.length != 0) {
+        this.setUtilityData();
+      }
     });
   }
 
   ngOnDestroy() {
-    this.facilityMetersSub.unsubscribe();
+    this.accountMeterDataSub.unsubscribe();
+    this.selectedFacilitySub.unsubscribe();
+    this.accountMetersSub.unsubscribe();
   }
 
   ngAfterViewInit() {
-    this.drawChart();
+    this.setUtilityData();
   }
+
   setUtilityData() {
-    if (this.facilityMeters) {
+    if (this.facilityMeters && this.facilityMeters.length != 0 && this.accountMeters && this.accountMeters.length != 0) {
       this.electricityData = this.getDataByUtility('Electricity', this.facilityMeters);
       this.naturalGasData = this.getDataByUtility('Natural Gas', this.facilityMeters);
       this.otherFuelsData = this.getDataByUtility('Other Fuels', this.facilityMeters);
@@ -57,16 +84,15 @@ export class FacilityStackedAreaChartComponent implements OnInit {
   }
 
 
-
   drawChart() {
-    if (this.stackedAreaChart) {
+    if (this.utilityBarChart) {
       let traceData = new Array();
       if (this.electricityData.length != 0) {
         let trace = {
           x: this.electricityData.map(data => { return data.time }),
           y: this.electricityData.map(data => { return data.energyCost }),
           name: 'Electricity',
-          stackgroup: 'one'
+          type: 'bar'
         }
         traceData.push(trace);
       }
@@ -75,7 +101,7 @@ export class FacilityStackedAreaChartComponent implements OnInit {
           x: this.naturalGasData.map(data => { return data.time }),
           y: this.naturalGasData.map(data => { return data.energyCost }),
           name: 'Natural Gas',
-          stackgroup: 'one'
+          type: 'bar'
         };
         traceData.push(trace);
       }
@@ -84,7 +110,7 @@ export class FacilityStackedAreaChartComponent implements OnInit {
           x: this.otherFuelsData.map(data => { return data.time }),
           y: this.otherFuelsData.map(data => { return data.energyCost }),
           name: 'Other Fuels',
-          stackgroup: 'one'
+          type: 'bar'
         };
         traceData.push(trace);
       }
@@ -93,7 +119,7 @@ export class FacilityStackedAreaChartComponent implements OnInit {
           x: this.waterData.map(data => { return data.time }),
           y: this.waterData.map(data => { return data.energyCost }),
           name: 'Water',
-          stackgroup: 'one'
+          type: 'bar'
         };
         traceData.push(trace);
       }
@@ -102,7 +128,7 @@ export class FacilityStackedAreaChartComponent implements OnInit {
           x: this.wasteWaterData.map(data => { return data.time }),
           y: this.wasteWaterData.map(data => { return data.energyCost }),
           name: 'Waste Water',
-          stackgroup: 'one'
+          type: 'bar'
         };
         traceData.push(trace);
       }
@@ -111,39 +137,46 @@ export class FacilityStackedAreaChartComponent implements OnInit {
           x: this.otherUtilityData.map(data => { return data.time }),
           y: this.otherUtilityData.map(data => { return data.energyCost }),
           name: 'Other Utility',
-          stackgroup: 'one'
+          type: 'bar'
         };
         traceData.push(trace);
       }
+      let xAxisTitle: string = 'Year';
+      if (this.sumByMonth) {
+        xAxisTitle = 'Month';
+      }
+
+
       var layout = {
         barmode: 'group',
-        title: {
-          text: 'Utility Costs',
-          font: {
-            size: 24
-          },
-        },
+        // title: {
+        //   text: 'Utility Costs',
+        //   font: {
+        //     size: 24
+        //   },
+        // },
         xaxis: {
-          autotick: false,
-          title: {
-            text: 'Year',
-            font: {
-              size: 18
-            },
-          },
+          // title: {
+          //   text: xAxisTitle,
+          //   font: {
+          //     size: 18
+          //   },
+          // },
         },
         yaxis: {
-          title: {
-            text: 'Energy Cost',
-            font: {
-              size: 18
-            },
-          },
+          // title: {
+          //   text: 'Energy Cost',
+          //   font: {
+          //     size: 18
+          //   },
+          // },
           hoverformat: '$,.2f'
-        }
+        },
+        margin: { r: 0, t: 50 }
       };
       var config = { responsive: true };
-      this.plotlyService.newPlot(this.stackedAreaChart.nativeElement, traceData, layout, config);
+
+      this.plotlyService.newPlot(this.utilityBarChart.nativeElement, traceData, layout, config);
     }
   }
