+++ src/app/dashboard/facility-overview/facility-overview.component.html
@@ -26,9 +26,26 @@
     </div>
 </div>
 
+
 <div class="row">
     <div class="col plot">
-        <!-- <p>Energy Usage and Cost Per Month (2020)</p> -->
+        <p>Monthly Energy Costs</p>
+        <app-facility-heat-map></app-facility-heat-map>
+    </div>
+</div>
+
+
+<div class="row">
+    <div class="col plot">
+        <p>Annual Utility Costs</p>
         <app-facility-bar-chart></app-facility-bar-chart>
+    </div>
+</div>
+
+
+<div class="row">
+    <div class="col plot">
+        <p>Annual Utility Costs</p>
+        <app-facility-stacked-area-chart></app-facility-stacked-area-chart>
     </div>
 </div>
