+++ src/app/dashboard/facility-overview/facility-stacked-area-chart/facility-stacked-area-chart.component.ts
@@ -2,53 +2,60 @@
 import { Component, ElementRef, OnInit, ViewChild } from '@angular/core';
 import { PlotlyService } from 'angular-plotly.js';
 import { Subscription } from 'rxjs';
+import { FacilitydbService } from 'src/app/indexedDB/facility-db.service';
 import { UtilityMeterdbService } from 'src/app/indexedDB/utilityMeter-db.service';
 import { UtilityMeterDatadbService } from 'src/app/indexedDB/utilityMeterData-db.service';
-import { IdbUtilityMeter } from 'src/app/models/idb';
-import { VisualizationService } from '../visualization.service';
+import { IdbFacility, IdbUtilityMeter } from 'src/app/models/idb';
+import { CalanderizationService } from '../../../utility/calanderization/calanderization.service';
+import { VisualizationService } from '../../../utility/visualization/visualization.service';
 
 @Component({
-  selector: 'app-facility-bar-chart',
-  templateUrl: './facility-bar-chart.component.html',
-  styleUrls: ['./facility-bar-chart.component.css']
+  selector: 'app-facility-stacked-area-chart',
+  templateUrl: './facility-stacked-area-chart.component.html',
+  styleUrls: ['./facility-stacked-area-chart.component.css']
 })
-export class FacilityBarChartComponent implements OnInit {
-  /*Grouped bar chart
-    Group by utility
-    Bars are energy usage and cost by time
-    Month or year (user selection)
-  */
+export class FacilityStackedAreaChartComponent implements OnInit {
 
+  @ViewChild('stackedAreaChart', { static: false }) stackedAreaChart: ElementRef;
 
-  @ViewChild('utilityBarChart', { static: false }) utilityBarChart: ElementRef;
-
-  facilityMetersSub: Subscription;
-
-
+  facilityMeters: Array<IdbUtilityMeter>;
   electricityData: Array<{ time: string, energyUse: number, energyCost: number }>;
   naturalGasData: Array<{ time: string, energyUse: number, energyCost: number }>;
   otherFuelsData: Array<{ time: string, energyUse: number, energyCost: number }>;
   waterData: Array<{ time: string, energyUse: number, energyCost: number }>;
   wasteWaterData: Array<{ time: string, energyUse: number, energyCost: number }>;
   otherUtilityData: Array<{ time: string, energyUse: number, energyCost: number }>;
-  facilityMeters: Array<IdbUtilityMeter>;
   sumByMonth: boolean = false;
   removeIncompleteYears: boolean = true;
+
+  selectedFacility: IdbFacility;
+  selectedFacilitySub: Subscription;
+  accountMeterDataSub: Subscription;
+  accountMeters: Array<IdbUtilityMeter>;
+  accountMetersSub: Subscription;
   constructor(private plotlyService: PlotlyService, private utilityMeterDbService: UtilityMeterdbService,
     private utilityMeterDataDbService: UtilityMeterDatadbService, private vizualizationService: VisualizationService) { }
 
   ngOnInit(): void {
-    this.facilityMetersSub = this.utilityMeterDataDbService.accountMeterData.subscribe(accountMeterData => {
+    this.accountMetersSub = this.utilityMeterDbService.accountMeters.subscribe(accountMeters => {
+      this.accountMeters = accountMeters;
+      this.setUtilityData();
+    });
+    this.selectedFacilitySub = this.utilityMeterDbService.facilityMeters.subscribe(facilityMeters => {
+      this.facilityMeters = JSON.parse(JSON.stringify(facilityMeters));
+      this.setUtilityData();
+    });
+    this.accountMeterDataSub = this.utilityMeterDataDbService.accountMeterData.subscribe(accountMeterData => {
       if (accountMeterData && accountMeterData.length != 0) {
-        let facilityMeters: Array<IdbUtilityMeter> = this.utilityMeterDbService.facilityMeters.getValue();
-        this.facilityMeters = JSON.parse(JSON.stringify(facilityMeters))
         this.setUtilityData();
       }
     });
   }
 
   ngOnDestroy() {
-    this.facilityMetersSub.unsubscribe();
+    this.accountMeterDataSub.unsubscribe();
+    this.selectedFacilitySub.unsubscribe();
+    this.accountMetersSub.unsubscribe();
   }
 
   ngAfterViewInit() {
@@ -56,7 +63,7 @@ export class FacilityBarChartComponent implements OnInit {
   }
 
   setUtilityData() {
-    if (this.facilityMeters) {
+    if (this.facilityMeters && this.facilityMeters.length != 0 && this.accountMeters && this.accountMeters.length != 0) {
       this.electricityData = this.getDataByUtility('Electricity', this.facilityMeters);
       this.naturalGasData = this.getDataByUtility('Natural Gas', this.facilityMeters);
       this.otherFuelsData = this.getDataByUtility('Other Fuels', this.facilityMeters);
@@ -68,15 +75,16 @@ export class FacilityBarChartComponent implements OnInit {
   }
 
 
+
   drawChart() {
-    if (this.utilityBarChart) {
+    if (this.stackedAreaChart) {
       let traceData = new Array();
       if (this.electricityData.length != 0) {
         let trace = {
           x: this.electricityData.map(data => { return data.time }),
           y: this.electricityData.map(data => { return data.energyCost }),
           name: 'Electricity',
-          type: 'bar'
+          stackgroup: 'one'
         }
         traceData.push(trace);
       }
@@ -85,7 +93,7 @@ export class FacilityBarChartComponent implements OnInit {
           x: this.naturalGasData.map(data => { return data.time }),
           y: this.naturalGasData.map(data => { return data.energyCost }),
           name: 'Natural Gas',
-          type: 'bar'
+          stackgroup: 'one'
         };
         traceData.push(trace);
       }
@@ -94,7 +102,7 @@ export class FacilityBarChartComponent implements OnInit {
           x: this.otherFuelsData.map(data => { return data.time }),
           y: this.otherFuelsData.map(data => { return data.energyCost }),
           name: 'Other Fuels',
-          type: 'bar'
+          stackgroup: 'one'
         };
         traceData.push(trace);
       }
@@ -103,7 +111,7 @@ export class FacilityBarChartComponent implements OnInit {
           x: this.waterData.map(data => { return data.time }),
           y: this.waterData.map(data => { return data.energyCost }),
           name: 'Water',
-          type: 'bar'
+          stackgroup: 'one'
         };
         traceData.push(trace);
       }
@@ -112,7 +120,7 @@ export class FacilityBarChartComponent implements OnInit {
           x: this.wasteWaterData.map(data => { return data.time }),
           y: this.wasteWaterData.map(data => { return data.energyCost }),
           name: 'Waste Water',
-          type: 'bar'
+          stackgroup: 'one'
         };
         traceData.push(trace);
       }
@@ -121,16 +129,10 @@ export class FacilityBarChartComponent implements OnInit {
           x: this.otherUtilityData.map(data => { return data.time }),
           y: this.otherUtilityData.map(data => { return data.energyCost }),
           name: 'Other Utility',
-          type: 'bar'
+          stackgroup: 'one'
         };
         traceData.push(trace);
       }
-      let xAxisTitle: string = 'Year';
-      if (this.sumByMonth) {
-        xAxisTitle = 'Month';
-      }
-
-
       var layout = {
         barmode: 'group',
         // title: {
@@ -140,27 +142,26 @@ export class FacilityBarChartComponent implements OnInit {
         //   },
         // },
         xaxis: {
-          // title: {
-          //   text: xAxisTitle,
-          //   font: {
-          //     size: 18
-          //   },
-          // },
-        },
-        yaxis: {
+          autotick: false,
           title: {
-            text: 'Energy Cost',
+            text: 'Year',
             font: {
               size: 18
             },
           },
-          hoverformat: '$,.2f'
         },
-        margin: { r: 0, t: 50 }
+        yaxis: {
+          // title: {
+          //   text: 'Energy Cost',
+          //   font: {
+          //     size: 18
+          //   },
+          // },
+          hoverformat: '$,.2f'
+        }
       };
       var config = { responsive: true };
-
-      this.plotlyService.newPlot(this.utilityBarChart.nativeElement, traceData, layout, config);
+      this.plotlyService.newPlot(this.stackedAreaChart.nativeElement, traceData, layout, config);
     }
   }
