+++ src/app/indexedDB/utilityMeter-db.service.ts
@@ -15,13 +15,13 @@ export class UtilityMeterdbService {
     constructor(private dbService: NgxIndexedDBService, private facilityDbService: FacilitydbService, private accountdbService: AccountdbService) {
         this.facilityMeters = new BehaviorSubject<Array<IdbUtilityMeter>>(new Array());
         this.accountMeters = new BehaviorSubject<Array<IdbUtilityMeter>>(new Array());
+        this.accountdbService.selectedAccount.subscribe(() => {
+            this.setAccountMeters();
+        })
         this.facilityDbService.selectedFacility.subscribe(() => {
             this.setFacilityMeters();
         });
 
-        this.accountdbService.selectedAccount.subscribe(() => {
-            this.setAccountMeters();
-        })
     }
 
     setFacilityMeters() {
