--- src/app/utility/visualization/visualization.component.html
@@ -1,17 +1,0 @@
-<div class="row">
-    <div class="col p-3">
-        <app-facility-bar-chart></app-facility-bar-chart>
-    </div>
-</div>
-
-<div class="row">
-    <div class="col p-3">
-        <app-facility-stacked-area-chart></app-facility-stacked-area-chart>
-    </div>
-</div>
-
-<div class="row">
-    <div class="col p-3">
-        <app-facility-heat-map></app-facility-heat-map>
-    </div>
-</div>
